Mpre structure changes
