+++ tests/components/assets/docker-compose.yaml
@@ -1,0 +1,47 @@
+version: "3.9" # optional since v1.27.0
+services:
+  postgres:
+    image: postgres:latest
+    ports:
+      - "5432:5432"
+    volumes:
+      - .:/mnt/db/postgres
+    environment:
+      - POSTGRES_PASSWORD=postgres
+    healthcheck:
+      test: ["CMD-SHELL", "pg_isready -U postgres -d postgres"]
+      interval: 30s
+      timeout: 60s
+      retries: 5
+      start_period: 80s
+  fuel-node:
+    image: local/composable-indexer-test:latest
+    command: >
+      bash -c "RUST_BACKTRACE=full ./fuel-node --wallet-path ./wallet.json --bin-path ./ping.bin"
+    ports:
+      - "4000:4000"
+    volumes:
+      - .:/mnt/db/fuel-indexer
+  web-api:
+    image: local/composable-indexer-test:latest
+    command: >
+      bash -c "RUST_BACKTRACE=full ./web-api --fuel-node-host fuel-node --fuel-node-port 4000 --wallet-path ./wallet.json --bin-path ./ping.bin"
+    ports:
+      - "8000:8000"
+    volumes:
+      - .:/mnt/db/fuel-indexer
+    depends_on:
+      - fuel-node
+  fuel-indexer:
+    image: local/composable-indexer-test:latest
+    # command: bash -c "sleep 2 && RUST_BACKTRACE=full ./fuel-indexer --fuel-node-host fuel-node --fuel-node-port 4000 --postgres-host postgres --postgres-password postgres --manifest manifest.yaml"
+    command: bash -c "sleep 2 && RUST_BACKTRACE=full ./fuel-indexer --fuel-node-host fuel-node --fuel-node-port 4000 --postgres-host postgres --postgres-password postgres --graphql-api-host 0.0.0.0"
+    ports:
+      - "29987:29987"
+    volumes:
+      - .:/mnt/db/fuel-indexer
+    depends_on:
+      - fuel-node
+      - web-api
+      - postgres
+
