+++ tests/components/assets/manifest/composability_test.yaml
@@ -1,0 +1,9 @@
+namespace: composability_test
+graphql_schema: composable-indexer-lib/schema/schema.graphql
+identifier: index1
+module:
+  wasm:
+    composable-indexer-lib/composable_indexer_lib.wasm
+handlers:
+  - event: LogData
+    handler: function_one
