+++ tests/components/assets/schema/composability_test.graphql
@@ -1,0 +1,15 @@
+schema {
+    query: QueryRoot
+}
+
+type QueryRoot {
+    message: Message
+}
+
+type Message {
+    id: ID!
+    ping: UInt8!
+    pong: UInt8!
+    message: Bytes32!
+}
+
