+++ tests/components/assets/schema/simple_wasm.graphql
