+++ tests/components/assets/wallet.json
