+++ tests/components/contracts/ping/Forc.lock
@@ -1,0 +1,14 @@
+[[package]]
+name = 'core'
+source = 'path+from-root-B4D60689AE851FD3'
+dependencies = []
+
+[[package]]
+name = 'ping'
+source = 'root'
+dependencies = ['std']
+
+[[package]]
+name = 'std'
+source = 'git+https://github.com/fuellabs/sway?tag=v0.19.2#6808861389966f99887f71476918a562a9edd90e'
+dependencies = ['core']
