+++ tests/components/contracts/ping/Forc.toml
@@ -1,0 +1,7 @@
+[project]
+authors = ["Rashad Alston"]
+entry = "main.sw"
+license = "Apache-2.0"
+name = "ping"
+
+[dependencies]
