+++ tests/components/contracts/simple_wasm/.gitignore
@@ -1,0 +1,2 @@
+out
+target
