+++ tests/components/contracts/simple_wasm/Forc.lock
@@ -1,0 +1,14 @@
+[[package]]
+name = 'contracts'
+source = 'root'
+dependencies = ['std']
+
+[[package]]
+name = 'core'
+source = 'path+from-root-4DD64E4644BF9794'
+dependencies = []
+
+[[package]]
+name = 'std'
+source = 'git+https://github.com/fuellabs/sway?tag=v0.21.0#32b5b142b54d3d38ae1f7df69f465138e86de82d'
+dependencies = ['core']
