+++ tests/components/contracts/simple_wasm/Forc.toml
@@ -1,0 +1,7 @@
+[project]
+authors = ["TJ"]
+entry = "main.sw"
+license = "Apache-2.0"
+name = "contracts"
+
+[dependencies]
