--- tests/components/data/manifest.yaml
@@ -1,13 +1,0 @@
----
-namespace: test_namespace
-graphql_schema: ./tests/test_data/schema.graphql
-identifier: index1
-module:
-  wasm:
-    ./tests/test_data/simple_wasm.wasm
-handlers:
-  - event: an_event_name
-    handler: function_one
-  - event: another_event_name
-    handler: function_two
-test_events: []
